pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Shane Nolan <shanemichealnolan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
typer = {extras = ["all"], version = "^0.3.2"}
mysql-connector-python = "^8.0.26"
pydantic = "^1.8.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
wemake-python-styleguide = "^0.15.3"
mypy = "^0.910"
nitpick = "^0.27.0"
pytest-mock = "^3.6.1"

[tool.nitpick]
style = "https://raw.githubusercontent.com/wemake-services/wemake-python-styleguide/master/styles/nitpick-style-wemake.toml"

[tool.poetry.scripts]
888 = "app.cli:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
